Update testing.md user stories

+++ products/views.py
@@ -177,7 +177,11 @@ def edit_product(request, product_id):
 @login_required
 def delete_product(request, product_id):
     """ Delete a product from the store """
-    product = get_object_or_404(Product, pk=product_id)
-    product.delete()
-    messages.success(request, 'Product deleted')
-    return redirect(reverse('products'))
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry only site admin can delete products')
+        return redirect(reverse('home'))
+    else:
+        product = get_object_or_404(Product, pk=product_id)
+        product.delete()
+        messages.success(request, 'Product deleted')
+        return redirect(reverse('products'))
